SelectConnection and logging
Made necessary changes to work with Select connection, which allows the
connection to stay alive indefinitely.
This is after I ran into problems with the blocking one, which usually
dies after ~10 minutes.

Added logging to the class, this is much needed for debugging on remote
servers.

Never seen the onClose method getting called, probably because the IO
Loop of the select connection avoids this problem.